Add max length constraint to C octet string reading
Add a max length parameter when reading C octet strings to ensure that we don't read more data than the protocol allows for a field. Potentially we should be harder on bad data and reject the packet if the string is missing a null terminator in the proper length range but for now we will silently stop reading at the end of the field.