Bug Fix: Correct the definition of a TLV
According to section 3.2.4.1 of the SMPP v3.4 specification, the tag and length of a TLV are 2 octets in size always and so should be stored as u16s rather than a String and u32 respectively. Also remove the null that was being added between the tag and length when we incorrectly thought that the tag was a null terminated string.